+++ src/components/App.js
@@ -1,6 +1,5 @@
 import React, { Component } from 'react';
 import { Dimensions, Platform } from 'react-native';
-import ErrorUtils from 'ErrorUtils';
 import {
   Actions,
   Router,
@@ -9,6 +8,7 @@ import {
 } from 'react-native-router-flux';
 import AppProps from './AppProps';
 import FCM from 'react-native-fcm';
+import LayoutAnimation from './Shared/LayoutAnimation';
 
 // Define reducer to manage scenes
 const reducerCreate = params=> {
@@ -22,45 +22,6 @@ const reducerCreate = params=> {
   };
 };
 
-export const animationStyle = (props) => {
-  const { layout, position, scene } = props;
-  const direction = (scene.navigationState && scene.navigationState.direction) ?
-        scene.navigationState.direction : 'horizontal';
-  const index = scene.index;
-  const inputRange = [index - 1, index, index + 1];
-  const width = layout.initWidth;
-  const height = layout.initHeight;
-
-  const opacity = position.interpolate({
-    inputRange,
-    outputRange: [0, 1, 0.5],
-  });
-
-  const scale = position.interpolate({
-    inputRange,
-    outputRange: [1, 1, 0.95],
-  });
-
-  let translateX = 0;
-  let translateY = 0;
-
-  switch (direction) {
-    case 'fade':
-      translateX = 0;
-      translateY = 0;
-      break;
-  }
-
-  return {
-        opacity,
-        transform: [
-            { scale },
-            { translateX },
-            { translateY },
-        ],
-      };
-};
-
 class App extends Component {
   constructor(props) {
     super(props);
@@ -69,13 +30,14 @@ class App extends Component {
   render() {
     let routerProp = {
       createReducer: reducerCreate,
+      animationStyle: LayoutAnimation.animationStyle,
       backAndroidHandler: () => this.backAndroidHandler(),
     };
 
     let Scenes = AppProps.sceneProps.map((props, key) => <Scene {...props} />);
 
     return (
-      <Router {...routerProp} animationStyle={animationStyle}>
+      <Router {...routerProp}>
         <Scene {...AppProps.rootProp}>
           {Scenes}
         </Scene>
